feat(RTOS): Implement function to initialize threads and test PendSV
Kernel_tcb_start() function initializes the private stack and stack
pointer for each thread. The unused memory addresses are filled with the
hex code 0xBAADF00DU to make it easier to check in the memory addresses
and test to make sure the memory allocation is correct.

Kernel_initialize() function is used to set the priorities for Systick
and PendSV to ensure PendSV will tail-chain and NOT preempt Systick. The
context switching of PendSV should only happen after an ISR has been
fully serviced.

The triggering of PendSV was tested manually by stepping through the
code and manually setting the PENDSVSET bit to 1 in the SCB_ICSR
register debugging window.

The kernel scheduler was implemented barebones to allow testing manually
by setting the address of the next_thread in the debugging view. This is
used to test to make sure the bit setting works properly in the code,
and the PendSV Handler is indeed tail-chained at the after the call to
Systick handler. The skeleton C code is also implemented in PendSV in
order to extract the compiler's generated ASM code to manually write
inline ASM for context switching functionality.