feat(context switch: Implement inline ASM to allow context switching
The PendSV_Handler has been implemented with inline assembly code to
allow for proper context switching between threads. There were a lot of
problems with the syntax and at the end of it, explicitly writing
__asm() for each line was much easier to debug because it lets you
single step each line of code. If you encase the whole thing in one
__asm() the compiler treats it as one line so you can't step through it.

The context switching was tested using manual thread switching in the
debugger.
1) Let the code run for a second
2) Set a breakpoint at the if statement inside kernel_scheduler()
3) Set a breakpoint at the beginning of PendSV Handler
4) Manually set the next_thread to &blinky1 or &blinky2 thread in
   variables view
5) Resume the code, DO NOT STEP because stepping auto disables
   interrupts
6) Now manually step through the ASM code and check to make sure the
   values being loaded into the register match the values you see in
   next_thread and current_thread
7) When the BX LR instruction runs, the code should return to the
   main_blinky1 or main_blinky2 thread and toggle the corresponding LED