feat(round robin): Add third blinky thread and refactor main while loop.
Start a third blinky to visually see how the context switching gives the
illusion of all 3 threads running concurrently.

The main while loop inside of main.c is redundant in an RTOS
implementation since the superloop will never be used. For programming
clarity, the superloop is replaced by a kernel_start() function to
improve the logic of the RTOS starting in the program.