feat(blocking): Implement idle thread to be used for thread blocking
The idle thread is a core function of making a thread blocking
implementation work properly. There might be a possibility that all
threads are blocked at some point so the idle thread is the thread that
will ALWAYS be ready to run, without busy-waiting the processor so it
can process other tasks as they come in efficiently.