feat(demo): Modify code to demonstrate a live demo on the board.
The idea is to have the LEDs flash fast enough where the threads look
like they are executing and blocking/unblocking dynamically, but not too
fast where it looks like a rave and it would hurt the eyes of the
viewer.

Some quirky additions had to be turning off the green and blue led
manually right before a context switch in the scheduler because the blue LED thread is
the lowest priority and sometimes when its preempted by another higher
priority thread, it would happen in between the two toggle instructions.
Thus leading to a situation where visually it would look like the thread
is still running when its actually not.